[feature/hls_conv_hbus] Add HLS CONV2D engine on HBUS (#138)
- [x] [hw/units] Add and build HLS CONV2D sources in custom IP
- [x] [hw/units] Update units' fetch sources parallelism
- [x] [sw] Add host sw code for HLS IP
- [x] [hw/xilinx] Restore HBUS
- [x] [hw/xilinx] Update DDR channel notation to DDR4CH\<n\>
- [x] [hw/xilinx] Move DDR ch0 (DDR4CH0) on HBUS
- [x] [hw/xilinx] Add DDR ch1 (DDR4CH1) on MBUS
- [x] [hw/xilinx] Integrate HLS IP wrapper in SoC
- [x] [hw/xilinx] Add readback support for HPC profile
- [x] [hw/xilinx/scripts] Automatic PCIe BAR detection and fixes for
devmem-based scripts
- [x] [hw/xilinx/scripts] Update GDB script

# Test plan 

## HLS
-  test on Alveo u250 MicroblazeV_RV32
    - [x] custom_hls_conv_hbus
 - test on Alveo u250 CV32EP
    - [x] custom_hls_conv_hbus
- ~test on Alveo u250 IBEX~ 
> Ibex fails timing at 100MHz
-  test on Alveo u250 CV64A6
    - [x] custom_hls_conv_hbus
-  test on Alveo u250 MicroblazeV_RV64
    - [x] custom_hls_conv_hbus
 - ~test on Alveo u280 CV32EP~
    - ~[ ] custom_hls_conv_hbus~
> On Alveo u280, the HLS sw example returns wrong results. This might be
related to the difference in Vivado version. Further debug will be
necessary, but out of the scope of this PR.

## Regression
- test on Nexys MicroblazeV_RV32
    - [x] hello_world
    - [x] echo
- test on Nexys MicroblazeV_RV64
    - [x] hello_world
    - [x] echo
- test on Nexys CV32EP
    - [x] hello_world
    - [x] echo
    - [x] interrupts
- test on Nexys IBEX
    - [x] hello_world
    - [x] echo
    - [x] interrupts
-  test on Nexys CV64A6
    - [x] hello_world
    - [x] echo
    - [x] interrupts
-  test on Alveo u250 MicroblazeV_RV32
    - [x] hello_world
    - [x] echo
-  test on Alveo u250 MicroblazeV_RV64
    - [x] hello_world
    - [x] echo
- test on Alveo u250 CV32EP
    - [x] hello_world
    - [x] echo
    - [x] interrupts
- test on Alveo u250 IBEX
    - [x] hello_world
    - [x] echo
    - [x] interrupts
-  test on Alveo u250 CV64A6
    - [x] hello_world
    - [x] echo
    - [x] interrupts
 - test on Alveo u280 CV32EP
    - [x] hello_world
    - [x] echo
    - [x] interrupts

---------

+++ sw/SoC/lib/uninasoc/inc/uninasoc.h
@@ -30,9 +30,9 @@ enum{
 
 static inline void uninasoc_init()
 {
-    extern const volatile uint32_t _peripheral_UART_start;
+    extern const volatile uintptr_t _peripheral_UART_start;
     // TinyIO init
-    uint32_t uart_base_address = (uint32_t)&_peripheral_UART_start;
+    uintptr_t uart_base_address = (uintptr_t)&_peripheral_UART_start;
     tinyIO_init(uart_base_address);
 }
